+++ src/main/java/com/podzilla/courier/services/delivery_task/DeliveryTaskService.java
@@ -10,9 +10,9 @@ import com.podzilla.courier.models.DeliveryStatus;
 import com.podzilla.courier.models.DeliveryTask;
 import com.podzilla.courier.repositories.delivery_task.IDeliveryTaskRepository;
 import com.podzilla.courier.services.delivery_task.confirmation_strategy.DeliveryConfirmationStrategy;
-import com.podzilla.courier.services.delivery_task.publish_command.Command;
-import com.podzilla.courier.services.delivery_task.publish_command.StopPollingCommand;
-import com.podzilla.courier.services.delivery_task.publish_command.StartPollingCommand;
+import com.podzilla.courier.services.delivery_task.poll_command.Command;
+import com.podzilla.courier.services.delivery_task.poll_command.StopPollingCommand;
+import com.podzilla.courier.services.delivery_task.poll_command.StartPollingCommand;
 import com.podzilla.mq.EventPublisher;
 import com.podzilla.mq.events.OrderCancelledEvent;
 import com.podzilla.mq.events.OrderOutForDeliveryEvent;
