+++ src/main/java/com/podzilla/courier/services/delivery_task/poll_command/Command.java
@@ -1,0 +1,5 @@
+package com.podzilla.courier.services.delivery_task.poll_command;
+
+public interface Command {
+    void execute();
+}
