--- src/main/java/com/podzilla/courier/services/delivery_task/publish_command/Command.java
@@ -1,5 +1,0 @@
-package com.podzilla.courier.services.delivery_task.publish_command;
-
-public interface Command {
-    void execute();
-}
